% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_boundary}
\alias{sedgwick_boundary}
\title{Sedgwick reserve boundary.}
\format{A SpatialPolygons object.
\describe{
}}
\usage{
sedgwick_boundary
}
\description{
Spatial boundary of Sedgwick Reserve for making maps. WGS84 geographical projection.
}
\keyword{datasets}
