% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_DEM}
\alias{sedgwick_DEM}
\title{Sedgwick Elevation Layer}
\format{A RasterLayer object with 12 slots
\describe{
}}
\usage{
sedgwick_DEM
}
\description{
Digital Elevation Model for Sedgwick Reserve downloaded from USGS.
1/3 Arc-second Resolution (~10m). Geographical projection on GRS80 ellipsoid.
}
\keyword{datasets}
