% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_env}
\alias{sedgwick_env}
\title{Geolocation and environmental characteristics of 24 study sites at the UCSB Sedgwick Reserve.}
\format{A data frame with 24 rows and 30 variables:
\describe{
  \item{site}{Site code based on original site number used by G.K. and N.K.}
  \item{name}{site nickname}
  \item{lon}{decimal longitude}
  \item{lat}{decimal latitude}
  \item{ele}{elevation (m), determined from USGS DEM}
  \item{type}{sites either upper or lower, upper sites on serpentine soil}
  \item{microsite}{for upper sites, sites on hummocks are identified}
  \item{Tmax}{Average daily Tmax \ifelse{html}{\out{&#176;}}{\&deg;}C temperature for March. Recorded directly with iButton placed in PVC housing on soil surface at each site}
  \item{Tmin}{Average daily Tmax \ifelse{html}{\out{&#176;}}{\&deg;}C temperature for March. Recorded directly with iButton placed in PVC housing on soil surface at each site}
  \item{soil_organic}{soil organic matter (\%)}
  \item{pH}{soil pH}
  \item{CEC}{soil cation exchange capacity (meq per 100g)}
  \item{NH4_N}{soil ammonium concentration (ppm)}
  \item{Nitrate}{soil Nitrate (ppm)}
  \item{sand}{Soil sand content (\%)}
  \item{clay}{Soil clay content (\%)}
  \item{Ca}{soil Calcium (ppm)}
  \item{Mg}{soil Magnesium (ppm)}
  \item{K}{soil Potassium (ppm)}
  \item{light_above}{Ambient light (mmol per m\ifelse{html}{\out{<sup>3</sup>}}{\eqn{^3}} measured at 1.5 m above site at two locations at midday}
  \item{light_below}{Ambient light (mmol per m\ifelse{html}{\out{<sup>3</sup>}}{\eqn{^3}} measured at soil surface, below any vegetion at the site}
  \item{light_use}{vegetation light interception.  Calculated as (light_above - light_below)/light_above}
  \item{soil_depth}{Average soil depth (cm). Determined by driving metal rod into two locations in five plots at each site, n = 10}
  \item{soil_depth_sd}{Standard deviation (cm) of soil depth measurements at each site. n = 10 per site}
  \item{Apr_2016_GWC}{Soil water content from soils collected April 2016 (g/g)}
  \item{Apr_2017_GWC}{Soil water content from soils collected April 2017 (g/g)}
  \item{Jan_2018_GWC}{Soil water content from soils collected January 2018 (g/g)}
  \item{May_2016_GWC}{Soil water content from soils collected May 2016 (g/g)}
  \item{avg_soil_moisture}{Average of April and May soil moisture measurements}
}}
\usage{
sedgwick_env
}
\description{
A dataset describing the position, soil characteristics, light interception,
and surface temperature at 24 vegetation study sites at the UCSB Sedgwick Reserve.
Sites originally established by Nathan Kraft and Gaurav Kandlikar in 2016.
Soil characteristics are based on standard soil analysis conducted by A&L. Western
Agricultural labs on samples collected from common gardens at each site by
Gaurav Kandlikar. Soil depth and light interception data were collected by
Andrew Kleinhesselink.
}
\keyword{datasets}
