% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_ibutton}
\alias{sedgwick_ibutton}
\title{Sedgwick Surface Temperature Data}
\format{A dataframe with 6636 rows and 7 variables
\describe{
  \item{site}{Site code based on original site number used by G.K. and N.K.}
  \item{date}{Date of sample in "Y-m-d" format}
  \item{doy}{Day of year}
  \item{Tmin}{Minimum temperature \ifelse{html}{\out{&#176;}}{\&deg;}C recorded during that day}
  \item{Tmax}{Maximum temperature \ifelse{html}{\out{&#176;}}{\&deg;}C recorded during that day}
  \item{Tavg}{Average temperature \ifelse{html}{\out{&#176;}}{\&deg;}C.  Tavg \= Tmin + Tamx/2 }
  \item{daily_range}{Range between maximum and minimum temperature \ifelse{html}{\out{&#176;}}{\&deg;}C during day}
}}
\usage{
sedgwick_ibutton
}
\description{
Spring surface temperatures for each of 24 study sites.  Temperatures recorded with
iButton temperature loggers housed in short section of white PVC pipe placed on soil
surface at each site. Temperatures were logged every 4 hours and maximum and minimum
daily temperatures were calculated. Temperatures were recorded from February to June
2016 and then again in 2018.
}
\keyword{datasets}
