% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_soil_depth}
\alias{sedgwick_soil_depth}
\title{Soil depth at each of the 24 sites.}
\format{A dataframe with 240 rows and 4 variables
\describe{
  \item{site}{Site code based on original site number used by G.K. and N.K.}
  \item{plot}{Plot number at each site (1-5)}
  \item{rep}{Two sampling locations located at opposite corners of each plot}
  \item{depth}{maximum depth (cm)}
}}
\usage{
sedgwick_soil_depth
}
\description{
Soil depth recorded at two locations within five plots (n=10) at each of the 24 site.
Soil depth was measured by driving a thin metal rod into the soil and recorded the
depth at which rock was hit and the rod stopped penetrating. Depth measured in the
spring of 2017 by A. Kleinhesselink.
}
\keyword{datasets}
