% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_soil_moisture}
\alias{sedgwick_soil_moisture}
\title{Spring moisture at each of the 24 sites.}
\format{A dataframe with 96 rows and 3 variables
\describe{
  \item{site}{Site code based on original site number used by G.K. and N.K.}
  \item{date}{Date of sample in "Y-m-d" format}
  \item{GWC}{Soil moisture expressed as fraction of dry weight}
}}
\usage{
sedgwick_soil_moisture
}
\description{
Soil moisture was determined gravimetrically in the spring of 2016, 2017 and 2018
at each of the 24 sites. Shallow (<10 cm) soil was collected from multiple locations
at each site.  These were homogenized and then weighed to get a wet weight.  They were
then dried at 60 \ifelse{html}{\out{&#176;}}{\&deg;}C for 72 hours and then re-weighed
to get dry weight.  Soil moisture by mass was recorded as (wet-dry)/dry weight.
}
\keyword{datasets}
