% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_veg_height}
\alias{sedgwick_veg_height}
\title{Vegetation height at each of the 24 sites.}
\format{A dataframe with 240 rows and 4 variables
\describe{
  \item{site}{Site code based on original site number used by G.K. and N.K.}
  \item{plot}{Plot number at each site (1-5)}
  \item{veg_height}{maximum vegetation height (cm)}
}}
\usage{
sedgwick_veg_height
}
\description{
Vegetation height was recorded in two places in each plot at each site.  Height was
determined by dropping a ruler in two opposite corners of the plot and recording
the height of the vegetation nearest to the ruler.
}
\keyword{datasets}
